ramips: add YTXC OEM AP support
Panel AP
Soc: MediaTek MT7628
Flash: 8MB
Memory: 128MB DDR2
USB 2.0: x 1
WIFI: 2.4G

Upgrade OpenWrt firmware:

Hardware Requirements:
Prepare a Windows-based computer and a PoE switch.
Connect the PoE switch's power supply port to the back panel port of the AP using an Ethernet cable. Connect the computer's Ethernet cable to the front panel port of the AP.
Ensure that the uplink port of the PoE switch is already connected to the network.

Download breed-mt7628-hiwifi-hc5661a.bin and transfer it to your local HTTP server.
Then open a web browser and log in to 192.168.168.1.
The web address will change to something like: http://192.168.168.1/cgi-bin/luci/;stok=???#/home/status.
Replace #/home/status with /admin/system/admin, and you will enter the OpenWrt LuCI page.
In the Status - Kernel Log page, verify that the CPU is MT7628.
Find the Custom Commands section in the system, add the following commands (check the "Do not verify" option):

Command 1:
wget http://your_http_server_ip/breed-mt7628-hiwifi-hc5661a.bin -O /tmp/breed.bin

Command 2:
mtd write /tmp/breed.bin Bootloader

Then save and apply the changes.
Go to the dashboard and execute Command 1 and Command 2 in sequence.
At this point, unplug the Ethernet cable from the uplink port of the PoE switch.
Then unplug the Ethernet cable from the front panel of the AP (the one connected directly to the computer).
Connect this cable to the uplink port of the PoE switch.
Then run breedenter on the computer and simultaneously press the POWER switch on the front panel of the AP to restart it.
This will allow you to successfully access the Breed page and flash the firmware.