ramips: IU-01W: Use unused GPIO pins on the PCB as the I2C
SDA:GPIO 0
SCL:GPIO 10

Test with SSD1306 OLED

Select the side of the PCB where the LED is located, and remove the SMD resistor closest to the TTL header in the top-left corner. This is ‌GPIO 0‌.
Flip the PCB over, locate the ‌78489 chip‌ beside the TTL's GND hole, and remove the SMD resistor below its ‌Pin 1‌. This is ‌GPIO 10‌.
Using ‌GPIO 0‌ and ‌GPIO 10‌ will enable proper operation of the I2C bus.